feat: 프로필 CRUD API 업데이트(질문 템플릿, 유형 테스트 반영)
- MyProfileResponse 수정
변경사항 반영
age, region, job, cosmicType, ChoiceTemplate(객관식), ShortTemplate(주관식) 추가
keyword, topic 제거

- ProfileService toMyProfileResponseDto 메서드 인자 수정 및 프로필 CRUD 메서드 조회 로직 추가
업데이트 된 MyProfileResponse 리턴하도록 변경

- QuestionService 추가
질문 템플릿, 유형 테스트 관련 엔티티 처리 서비스 계층

+++ src/main/java/org/swyp/com/backend/profile/service/ProfileService.java
@@ -10,14 +10,24 @@ import org.swyp.com.backend.global.enumeration.InterestType;
 import org.swyp.com.backend.global.exception.BusinessException;
 import org.swyp.com.backend.profile.domain.Interest;
 import org.swyp.com.backend.profile.domain.Profile;
+import org.swyp.com.backend.profile.domain.ProfileChoiceTemplate;
 import org.swyp.com.backend.profile.domain.ProfileInterest;
+import org.swyp.com.backend.profile.domain.ProfileShortTemplate;
 import org.swyp.com.backend.profile.domain.repository.InterestRepository;
+import org.swyp.com.backend.profile.domain.repository.ProfileChoiceTemplateRepository;
 import org.swyp.com.backend.profile.domain.repository.ProfileInterestRepository;
 import org.swyp.com.backend.profile.domain.repository.ProfileRepository;
+import org.swyp.com.backend.profile.domain.repository.ProfileShortTemplateRepository;
 import org.swyp.com.backend.profile.dto.MyProfileResponse;
 import org.swyp.com.backend.profile.dto.ProfileRegisterRequest;
 import org.swyp.com.backend.profile.dto.ProfileResponse;
 import org.swyp.com.backend.profile.dto.ProfileUpdateRequest;
+import org.swyp.com.backend.question.domain.MultipleChoiceAnswer;
+import org.swyp.com.backend.question.domain.MultipleChoiceQuestion;
+import org.swyp.com.backend.question.domain.ShortAnswerQuestion;
+import org.swyp.com.backend.question.dto.ChoiceTemplate;
+import org.swyp.com.backend.question.dto.ShortTemplate;
+import org.swyp.com.backend.question.service.QuestionService;
 import org.swyp.com.backend.user.domain.User;
 import org.swyp.com.backend.user.domain.repository.UserRepository;
 
@@ -25,10 +35,16 @@ import org.swyp.com.backend.user.domain.repository.UserRepository;
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class ProfileService {
+    private final QuestionService questionService;
+
     private final UserRepository userRepository;
     private final ProfileRepository profileRepository;
+
     private final InterestRepository interestRepository;
     private final ProfileInterestRepository profileInterestRepository;
+
+    private final ProfileChoiceTemplateRepository profileChoiceTemplateRepository;
+    private final ProfileShortTemplateRepository profileShortTemplateRepository;
 
     public MyProfileResponse getMyProfile(final Long userId) {
         User user = userRepository.findById(userId).orElseThrow(() ->
@@ -37,9 +53,11 @@ public class ProfileService {
         Profile profile = profileRepository.findByUser(user).orElseThrow(() ->
                 new BusinessException(HttpStatus.NOT_FOUND, "프로필 정보를 찾을 수 없습니다."));
 
-        List<ProfileInterest> interestList = profileInterestRepository.findByProfile(profile);
+        List<ProfileInterest> ProfileInterestList = profileInterestRepository.findByProfile(profile);
+        List<ProfileChoiceTemplate> profileChoiceList = profileChoiceTemplateRepository.findByProfile(profile);
+        List<ProfileShortTemplate> profileShortList = profileShortTemplateRepository.findByProfile(profile);
 
-        return toMyProfileResponseDto(profile, interestList);
+        return toMyProfileResponseDto(profile, ProfileInterestList, profileChoiceList, profileShortList);
     }
 
     public ProfileResponse getUserProfile(final Long uuid) {
@@ -54,16 +72,31 @@ public class ProfileService {
         if (profileRepository.findByUser(user).isPresent()) {
             throw new BusinessException(HttpStatus.CONFLICT, "이미 프로필을 생성하였습니다.");
         }
+
         Profile profile = Profile.createProfile(user, profileForm.nickname(), profileForm.imageKey(),
-                profileForm.gender(),
-                profileForm.bio(), profileForm.keyword(), profileForm.topic());
+                profileForm.gender(), profileForm.age(), profileForm.region(), profileForm.job(),
+                profileForm.bio(), profileForm.cosmicType());
 
         List<ProfileInterest> profileInterestList = toProfileInterestList(profile, profileForm.interests());
 
         profileRepository.save(profile);
         profileInterestRepository.saveAll(profileInterestList);
 
-        return toMyProfileResponseDto(profile, profileInterestList);
+        List<ProfileChoiceTemplate> profileChoiceTemplateList = new ArrayList<>();
+        List<ProfileShortTemplate> profileShortTemplateList = new ArrayList<>();
+
+        if (profileForm.choiceTemplate() != null) {
+            profileChoiceTemplateList = profileChoiceTemplateRepository.saveAll(
+                    questionService.toProfileChoiceList(profile, profileForm.choiceTemplate()));
+        }
+
+        if (profileForm.shortTemplate() != null) {
+            profileShortTemplateList = profileShortTemplateRepository.saveAll(
+                    questionService.toProfileShortList(profile, profileForm.shortTemplate()));
+        }
+
+        return toMyProfileResponseDto(profile, profileInterestList, profileChoiceTemplateList,
+                profileShortTemplateList);
     }
 
     @Transactional
@@ -74,16 +107,32 @@ public class ProfileService {
         Profile profile = profileRepository.findByUser(user).orElseThrow(() ->
                 new BusinessException(HttpStatus.NOT_FOUND, "프로필 정보를 찾을 수 없습니다."));
 
-        profile.updateProfile(profileForm.nickname(), profileForm.getImageKey(), profileForm.getBio(),
-                profileForm.getKeyword(), profileForm.getTopic());
+        profile.updateProfile(profileForm.nickname(), profileForm.imageKey(), profileForm.age(), profileForm.region(),
+                profileForm.job(), profileForm.bio(), profileForm.cosmicType());
 
-        profileInterestRepository.deleteByProfile(profile);
+        if (profileForm.interests() != null) {
+            profileInterestRepository.deleteByProfile(profile);
+            profileInterestRepository.saveAll(toProfileInterestList(profile, profileForm.interests()));
+        }
 
-        List<ProfileInterest> profileInterestList = toProfileInterestList(profile, profileForm.interests());
+        if (profileForm.choiceTemplate() != null) {
+            profileChoiceTemplateRepository.deleteByProfile(profile);
+            profileChoiceTemplateRepository.saveAll(
+                    questionService.toProfileChoiceList(profile, profileForm.choiceTemplate()));
+        }
 
-        profileInterestRepository.saveAll(profileInterestList);
+        if (profileForm.shortTemplate() != null) {
+            profileShortTemplateRepository.deleteByProfile(profile);
+            profileShortTemplateRepository.saveAll(
+                    questionService.toProfileShortList(profile, profileForm.shortTemplate()));
+        }
 
-        return toMyProfileResponseDto(profile, profileInterestList);
+        List<ProfileInterest> profileInterestList = profileInterestRepository.findByProfile(profile);
+        List<ProfileChoiceTemplate> profileChoiceTemplateList = profileChoiceTemplateRepository.findByProfile(profile);
+        List<ProfileShortTemplate> profileShortTemplateList = profileShortTemplateRepository.findByProfile(profile);
+
+        return toMyProfileResponseDto(profile, profileInterestList, profileChoiceTemplateList,
+                profileShortTemplateList);
     }
 
     @Transactional
@@ -95,19 +144,44 @@ public class ProfileService {
                 new BusinessException(HttpStatus.NOT_FOUND, "프로필 정보를 찾을 수 없습니다."));
 
         profileInterestRepository.deleteByProfile(profile);
+        profileChoiceTemplateRepository.deleteByProfile(profile);
+        profileShortTemplateRepository.deleteByProfile(profile);
         profileRepository.delete(profile);
     }
 
-    private MyProfileResponse toMyProfileResponseDto(Profile profile, List<ProfileInterest> interestList) {
+    private MyProfileResponse toMyProfileResponseDto(Profile profile, List<ProfileInterest> interestList,
+                                                     List<ProfileChoiceTemplate> profileChoiceTemplateList,
+                                                     List<ProfileShortTemplate> profileShortTemplateList) {
+
         List<InterestType> interestTypeList = new ArrayList<>();
+        List<ChoiceTemplate> choiceTemplateList = new ArrayList<>();
+        List<ShortTemplate> shortTemplateList = new ArrayList<>();
 
         for (ProfileInterest interest : interestList) {
             interestTypeList.add(interest.getInterest().getType());
         }
 
+        for (ProfileChoiceTemplate profileChoice : profileChoiceTemplateList) {
+            choiceTemplateList.add(
+                    toChoiceTemplate(profileChoice.getAnswer().getQuestion(), profileChoice.getAnswer()));
+        }
+
+        for (ProfileShortTemplate profileShort : profileShortTemplateList) {
+            shortTemplateList.add(toShortTemplate(profileShort.getQuestion(), profileShort.getAnswer()));
+        }
+
         return new MyProfileResponse(profile.getId(), profile.getNickname(),
-                profile.getImageKey(), profile.getGender(), profile.getBio(), profile.getKeyword(), profile.getTopic(),
-                interestTypeList);
+                profile.getImageKey(), profile.getGender(), profile.getAge(), profile.getRegion(), profile.getJob(),
+                interestTypeList, profile.getBio(), profile.getCosmic(), choiceTemplateList, shortTemplateList);
+    }
+
+    private ChoiceTemplate toChoiceTemplate(MultipleChoiceQuestion question, MultipleChoiceAnswer answer) {
+        return new ChoiceTemplate(question.getId(), question.getType(),
+                question.getContent(), answer.getAnswerId(), answer.getContent());
+    }
+
+    private ShortTemplate toShortTemplate(ShortAnswerQuestion question, String answer) {
+        return new ShortTemplate(question.getId(), question.getType(), question.getContent(), answer);
     }
 
     private List<ProfileInterest> toProfileInterestList(Profile profile, List<InterestType> interestTypeList) {
@@ -120,6 +194,4 @@ public class ProfileService {
 
         return profileInterestList;
     }
-
-
 }
